Improved macros slightly

+++ src/Atmel_Studio/Space-Buddies.h
@@ -57,6 +57,6 @@ void checkLeftButton(void);
 void checkButtons(void);
 void play(volatile uint8_t *pByte);
 void send_data(volatile uint8_t *pByte);
-int readCapacitivePin(volatile uint8_t* ddr, volatile uint8_t* port, volatile uint8_t* pin, uint8_t pinNumber);
+uint8_t readCapacitivePin(volatile uint8_t* ddr, volatile uint8_t* port, volatile uint8_t* pin, uint8_t pinNumber);
 void start_timer2(void);
 void stop_timer2(void);
